// src_frontend/api-client/models/MailingListRecipient.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface MailingListRecipient
 */
export interface MailingListRecipient {
  /**
   *
   * @type {string}
   * @memberof MailingListRecipient
   */
  address: string;
  /**
   *
   * @type {boolean}
   * @memberof MailingListRecipient
   */
  userConfirmed: boolean;
  /**
   *
   * @type {string}
   * @memberof MailingListRecipient
   */
  linkToMemberProfile: string | null;
}

/**
 * Check if a given object implements the MailingListRecipient interface.
 */
export function instanceOfMailingListRecipient(
  value: object,
): value is MailingListRecipient {
  if (!("address" in value) || value["address"] === undefined) return false;
  if (!("userConfirmed" in value) || value["userConfirmed"] === undefined)
    return false;
  if (
    !("linkToMemberProfile" in value) ||
    value["linkToMemberProfile"] === undefined
  )
    return false;
  return true;
}

export function MailingListRecipientFromJSON(json: any): MailingListRecipient {
  return MailingListRecipientFromJSONTyped(json, false);
}

export function MailingListRecipientFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): MailingListRecipient {
  if (json == null) {
    return json;
  }
  return {
    address: json["address"],
    userConfirmed: json["user_confirmed"],
    linkToMemberProfile: json["link_to_member_profile"],
  };
}

export function MailingListRecipientToJSON(json: any): MailingListRecipient {
  return MailingListRecipientToJSONTyped(json, false);
}

export function MailingListRecipientToJSONTyped(
  value?: MailingListRecipient | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    address: value["address"],
    user_confirmed: value["userConfirmed"],
    link_to_member_profile: value["linkToMemberProfile"],
  };
}